Change the Square to be a function component
In React, function components are a simpler way to write components that
only contain a render method and don't have their own state. Instead of
defining a class that extends React.Component, we can write a function
that takes props as input and returns what should be rendered.

Note: When we modified the Square to be a function component, we also
changed onClick={() => this.props.onClick()} to a shorter
onClick={props.onClick} (note the lack of parentheses on both sides).