Output state to the console
I've been trying to output the updated state of history length but I was
gettting an unexpected result. I learned that setState is asynchronous
and because of that, when I output to the console the state just after
its update, the value of the state would be outputted before its update.
There are two ways to solve this problem:

 - Use useEffect to display the state when the component renders

 - Use an optional function passed as argument to the setState function

I chose the last aproach, as you can see.